+++ src/ethereum_test_fixtures/blockchain.py
@@ -6,8 +6,7 @@ from functools import cached_property
 from typing import Annotated, Any, ClassVar, List, Literal, Tuple, Union, get_args, get_type_hints
 
 from ethereum import rlp as eth_rlp
-from ethereum.base_types import Uint
-from ethereum.crypto.hash import keccak256
+from ethereum_types.numeric import Uint
 from pydantic import AliasChoices, Field, PlainSerializer, computed_field
 
 from ethereum_test_base_types import (
@@ -24,7 +23,7 @@ from ethereum_test_base_types import (
     ZeroPaddedHexNumber,
 )
 from ethereum_test_exceptions import EngineAPIError, ExceptionInstanceOrList
-from ethereum_test_forks import Fork
+from ethereum_test_forks import Fork, Paris
 from ethereum_test_types.types import (
     AuthorizationTupleGeneric,
     ConsolidationRequest,
@@ -42,7 +41,6 @@ from ethereum_test_types.types import (
 )
 
 from .base import BaseFixture
-from .formats import FixtureFormats
 
 
 class HeaderForkRequirement(str):
@@ -182,7 +180,7 @@ class FixtureHeader(CamelModel):
         """
         Compute the RLP of the header
         """
-        return Hash(keccak256(self.rlp))
+        return self.rlp.keccak256()
 
 
 class FixtureExecutionPayload(CamelModel):
@@ -500,11 +498,12 @@ class Fixture(FixtureCommon):
     Cross-client specific blockchain test model use in JSON fixtures.
     """
 
+    fixture_format_name: ClassVar[str] = "blockchain_test"
+    description: ClassVar[str] = "Tests that generate a blockchain test fixture."
+
     genesis_rlp: Bytes = Field(..., alias="genesisRLP")
     blocks: List[FixtureBlock | InvalidFixtureBlock]
     seal_engine: Literal["NoProof"] = Field("NoProof")
-
-    format: ClassVar[FixtureFormats] = FixtureFormats.BLOCKCHAIN_TEST
 
 
 class EngineFixture(FixtureCommon):
@@ -512,7 +511,19 @@ class EngineFixture(FixtureCommon):
     Engine specific test fixture information.
     """
 
+    fixture_format_name: ClassVar[str] = "blockchain_test_engine"
+    description: ClassVar[
+        str
+    ] = "Tests that generate a blockchain test fixture in Engine API format."
+
     payloads: List[FixtureEngineNewPayload] = Field(..., alias="engineNewPayloads")
     sync_payload: FixtureEngineNewPayload | None = None
 
-    format: ClassVar[FixtureFormats] = FixtureFormats.BLOCKCHAIN_TEST_ENGINE
+    @classmethod
+    def supports_fork(cls, fork: Fork) -> bool:
+        """
+        Returns whether the fixture can be generated for the given fork.
+
+        The Engine API is available only on Paris and afterwards.
+        """
+        return fork >= Paris